Add Cart Changed: - Public: Index, CSS (product) - Controllers: Controller, Home and Product - Entities: Product - Services: Request and Product - Views: Product (index, edit, table) Added: - Public: CSS (cart) - Controllers: Cart - Services: Cart - Views: Cart

+++ src/entities/Product.php
@@ -66,7 +66,7 @@ class Product extends Entity
     {
         $this->price = 0;
         if (!empty($price) and is_numeric($price)) {
-            $this->price = (float)$price;
+            $this->price = (int)$price;
         }
     }
